requirements.txt
Flask
google-generativeai
manim
gunicorn # Production-ready web server
